+++ flopy4/utils.py
@@ -19,4 +19,8 @@ def flatten(l: Any):
 
 
 def to_path(value: Any) -> Optional[Path]:
+    """
+    Convert a value to a Path if it has a value,
+    otherwise return None.
+    """
     return Path(value) if value else None
