+++ pyproject.toml
@@ -97,6 +97,7 @@ select = [
     "I001", # isort - unsorted-imports
 ]
 ignore = [
+    "F841", # local variable is assigned to but never used
     "F821", # undefined name TODO FIXME
     "E722", # do not use bare `except`
     "E741", # ambiguous variable name
